Revert "Add logout function"
This reverts commit db6ba903a2c318bbf84c3efad297e944f6cac207.

+++ Sources/EtoileKit/Data/Lyrics.swift
@@ -10,6 +10,6 @@ import JellyfinAPI
 
 public struct Lyrics {
     public let lines: [LyricLine]?
-    public let offset: Int?
+    public let offset: Int64?
     public let synced: Bool?
 }
